Move next components into position. Remove next entrypoint.
