+++ packages/ui/react-ui-stack/package.json
@@ -20,11 +20,6 @@
         "require": "./dist/lib/node/testing/index.cjs",
         "default": "./dist/lib/node-esm/testing/index.mjs"
       }
-    },
-    "./next": {
-      "types": "./dist/types/src/next/index.d.ts",
-      "browser": "./dist/lib/browser/next/index.mjs",
-      "node": "./dist/lib/node-esm/next/index.mjs"
     }
   },
   "types": "dist/types/src/index.d.ts",
@@ -32,9 +27,6 @@
     "*": {
       "testing": [
         "dist/types/src/testing/index.d.ts"
-      ],
-      "next": [
-        "dist/types/src/next/index.d.ts"
       ]
     }
   },
